requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.15
psycopg2-binary>=2.9
gunicorn>=21.2
django-cors-headers>=4.3
python-decouple>=3.8
